Rename RN Sample application to sentry-react-native-sample
