+++ samples/react-native/android/app/build.gradle
@@ -129,9 +129,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdkVersion rootProject.ext.compileSdkVersion
 
-    namespace "com.samplenewarchitecture"
+    namespace "io.sentry.reactnative.sample"
     defaultConfig {
-        applicationId "com.samplenewarchitecture"
+        applicationId "io.sentry.reactnative.sample"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
