+++ samples/react-native/android/app/src/main/java/io/sentry/reactnative/sample/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.samplenewarchitecture
+package io.sentry.reactnative.sample
 
 import android.os.Bundle
 import com.facebook.react.ReactActivity
@@ -11,7 +11,7 @@ class MainActivity : ReactActivity() {
      * Returns the name of the main component registered from JavaScript. This is used to schedule
      * rendering of the component.
      */
-    override fun getMainComponentName(): String = "sampleNewArchitecture"
+    override fun getMainComponentName(): String = "sentry-react-native-sample"
 
     /**
      * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]
