+++ samples/react-native/android/app/src/main/java/io/sentry/reactnative/sample/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.samplenewarchitecture
+package io.sentry.reactnative.sample
 
 import android.app.Application
 import com.facebook.react.PackageList
