+++ samples/react-native/android/app/src/main/java/io/sentry/reactnative/sample/NativePlatformSampleModule.kt
@@ -1,4 +1,4 @@
-package com.samplenewarchitecture
+package io.sentry.reactnative.sample
 
 import com.facebook.fbreact.specs.NativePlatformSampleModuleSpec
 import com.facebook.react.bridge.ReactApplicationContext
