+++ samples/react-native/android/app/src/main/java/io/sentry/reactnative/sample/SamplePackage.java
@@ -1,4 +1,4 @@
-package com.samplenewarchitecture;
+package io.sentry.reactnative.sample;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
