+++ samples/react-native/android/app/src/main/java/io/sentry/reactnative/sample/TurboSamplePackage.kt
@@ -1,4 +1,4 @@
-package com.samplenewarchitecture
+package io.sentry.reactnative.sample
 
 import com.facebook.react.TurboReactPackage
 import com.facebook.react.bridge.NativeModule
