+++ samples/react-native/android/app/src/main/jni/OnLoad.cpp
@@ -100,7 +100,7 @@ JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM *vm, void *) {
 
 extern "C"
 JNIEXPORT void JNICALL
-Java_com_samplenewarchitecture_SamplePackage_crash(JNIEnv *env, jobject thiz) {
+Java_io_sentry_reactnative_sample_SamplePackage_crash(JNIEnv *env, jobject thiz) {
   char *ptr = 0;
   *ptr += 1;
 }
