+++ samples/react-native/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">sampleNewArchitecture</string>
+    <string name="app_name">Sentry RN</string>
 </resources>
