+++ samples/react-native/android/app/src/release/java/io/sentry/reactnative/sample/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.samplenewarchitecture;
+package io.sentry.reactnative.sample;
 import android.content.Context;
 import com.facebook.react.ReactInstanceManager;
 /**
