+++ samples/react-native/android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'sampleNewArchitecture'
+rootProject.name = 'sentry-react-native-sample'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 include ':app'
 includeBuild('../node_modules/@react-native/gradle-plugin')
