+++ samples/react-native/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "sampleNewArchitecture",
-  "displayName": "sampleNewArchitecture"
+  "name": "sentry-react-native-sample",
+  "displayName": "Sentry RN"
 }
