+++ samples/react-native/ios/sentryreactnativesample.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:sampleNewArchitecture.xcodeproj">
+      location = "group:sentryreactnativesample.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
