+++ samples/react-native/ios/sentryreactnativesample.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
