+++ samples/react-native/ios/sentryreactnativesample/AppDelegate.h
