+++ samples/react-native/ios/sentryreactnativesample/AppDelegate.mm
@@ -15,7 +15,7 @@
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
-  self.moduleName = @"sampleNewArchitecture";
+  self.moduleName = @"sentry-react-native-sample";
   // You can add your custom initial props in the dictionary below.
   // They will be passed down to the ViewController used by React
   self.initialProps = @{};
