+++ samples/react-native/ios/sentryreactnativesample/Images.xcassets/AppIcon.appiconset/Contents.json
