+++ samples/react-native/ios/sentryreactnativesample/Images.xcassets/Contents.json
