+++ samples/react-native/ios/sentryreactnativesample/Images.xcassets/ExampleBinaryData.dataset/Contents.json
