+++ samples/react-native/ios/sentryreactnativesample/Images.xcassets/ExampleBinaryData.dataset/logo_mini.png
