+++ samples/react-native/ios/sentryreactnativesample/NativePlatformSampleModule.h
