+++ samples/react-native/ios/sentryreactnativesample/NativePlatformSampleModule.mm
