+++ samples/react-native/ios/sentryreactnativesample/RCTAssetsModule.h
