+++ samples/react-native/ios/sentryreactnativesample/RCTAssetsModule.m
