+++ samples/react-native/ios/sentryreactnativesample/main.m
