+++ samples/react-native/ios/sentryreactnativesampleTests/Info.plist
