+++ samples/react-native/ios/sentryreactnativesampleTests/sentryreactnativesampleTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface sampleNewArchitectureTests : XCTestCase
+@interface sentryreactnativesampleTests : XCTestCase
 
 @end
 
-@implementation sampleNewArchitectureTests
+@implementation sentryreactnativesampleTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
