+++ samples/react-native/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "sentry-react-native-bare-sample",
+  "name": "sentry-react-native-sample",
   "version": "0.0.1",
   "private": true,
   "scripts": {
